Imported AWT renderer from HEAD in preparation for area generation and rendering experiments. Hacks to allow compilation.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/branches/FOP_0-20-0_Alt-Design@197440 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/render/awt/AWTFontMetrics.java
@@ -157,7 +157,7 @@ public class AWTFontMetrics {
      */
     public int getXHeight(String family, int style, int size) {
         setFont(family, style, size);
-        return (int)(FONT_FACTOR * xHeight);
+        return (FONT_FACTOR * xHeight);
     }
 
     /**
@@ -178,7 +178,7 @@ public class AWTFontMetrics {
         if (i <= 32) {
             w = (int)(1.4 * fmt.charWidth(i) * FONT_FACTOR);
         } else {
-            w = (int)(fmt.charWidth(i) * FONT_FACTOR);
+            w = (fmt.charWidth(i) * FONT_FACTOR);
         }
         return w;
     }
@@ -227,7 +227,7 @@ public class AWTFontMetrics {
             if ((this.style != style) || !this.family.equals(family)
                     || this.size != s) {
                 if (family.equals(this.family)) {
-                    f1 = f1.deriveFont(style, (float)s);
+                    f1 = f1.deriveFont(style, s);
                 } else {
                     f1 = new Font(family, style, s);
                 }
